pyproject.toml
[tool.poetry]
name = "agent"
version = "0.1.0"
description = ""
authors = ["Eser Ozvataf <eser@acikyazilim.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
phidata = "^2.7.6"
openai = "^1.59.3"
duckduckgo-search = "^7.2.0"
packaging = "^24.2"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
